added Markers on Map View

+++ app/src/main/java/at/jku/yourmemorylane/db/entities/Memory.kt
@@ -9,8 +9,8 @@ import java.sql.Date
 data class Memory(
     @ColumnInfo(name = "title") var title: String,
     @ColumnInfo(name = "date") var date: String,
-    @ColumnInfo(name = "longitude") var longitude: Int,
-    @ColumnInfo(name = "latitude") var latitude: Int
+    @ColumnInfo(name = "longitude") var longitude: Double,
+    @ColumnInfo(name = "latitude") var latitude: Double
 ){
     @PrimaryKey(autoGenerate = true) var id: Int =0
 }
